pyproject.toml
[project]
name = "openenv-ticket-triage"
version = "0.1.0"
description = "OpenEnv Round 1 environment for real-world support ticket triage."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "openenv-core==0.2.1",
  "fastapi==0.115.12",
  "uvicorn[standard]==0.34.1",
  "pydantic==2.11.3",
  "python-dotenv==1.1.0",
  "openai==2.30.0",
  "httpx==0.28.1",
  "PyYAML==6.0.2",
]

[project.scripts]
server = "server.app:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
